+++ null/net_null.c
@@ -276,7 +276,7 @@ void	NET_Config (qboolean multiplayer)
 }
 
 // sleeps msec or until net socket is ready
-void NET_Sleep(int msec)
+void NET_Sleep(double msec)
 {
 	if (!dedicated || !dedicated->value)
 		return; // we're not a server, just run full speed
@@ -288,8 +288,8 @@ void NET_Sleep(int msec)
 		FD_SET(ip_sockets[NS_SERVER], &fdset); // network socket
 		i = ip_sockets[NS_SERVER];
 	}
-	timeout.tv_sec = msec/1000;
-	timeout.tv_usec = (msec%1000)*1000;
+	timeout.tv_sec = (long)msec/1000;
+	timeout.tv_usec = ((long)msec%1000)*1000;
 	select(i+1, &fdset, NULL, NULL, &timeout);
 #endif
 }
